requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
py-cpuinfo==9.0.0
random2==1.0.1
Werkzeug==2.2.3
